Update DeployTheNCAgent.ps1
Updated the script to target the ServerEI2 version of the CustomerList API, as that version supports JWT authentication. As a result, the code that finds the registration token for the specified customerID has been modified (thanks Marc Andre Tanguay!). Also (thanks to Marc Andre) a line has been added that forces the script to use TLS 1.2.