Created script file
Originally posted @ https://github.com/oberjors/PublicShare.
This script creates a folder to hold the agent file for deployment on the launch of a provisioned image and the scripting necessary to execute on boot when returning to the gold state. *Should* function with PVS/MCS/Linked Clones/Horizon and the like. Limited testing was performed with PVS and VMWare cloning/template deployment.